saparate components modal
